Machine learning for face identification
Machine learning algorithm to learn which matrizes of grayscales belong to which person inside the pictures folder you created.
A facetrainer.yml data is created by running that program. This .yml data is then utilized in the face_identification.py code.
First execute this code before running the face_identification.py file! Otherwise an error message is going to be displayed.