Final full face identification and recognition
This is the full program which utilizes the recognition via the haarcascades and the .yml code which was created in the face_machine_learning.py to also display the name of the identified person above the head of the person in real time.
In order to execute this script properly you will have to fill the pictures folder with faces of persons and name folders and pictures correctly.
Afterwards run the face_machine_learning.py script, so that the facetrainer.yml ist created. Then you can finally run and enjoy this real time face recognition and identification application. Enjoy!